Read utf-8 data using utf-8-sig as encoding
This commit changes the way utf-8 data is read by **always** using the
utf-8-sig encoding, i.e. stripping away a BOM, if present. This has a
couple of ramifications:
- Since we do not use utf-8-sig for writing, this breaks full roundtripping
  for utf-8 data with BOM.
- There's no way to specify "utf-8 without BOM"; i.e. if file content starts
  with 0xfeff a BOM is required.

I still think the change is justified, because being able to properly read
CSV data as exported from Excel is really important. While this could also
be achieved by specifying a CSV dialect with utf-8-sig as encoding, this
solution doesn't translate well into CSVW, because utf-8-sig is only (if at
all) known in Python, and not a well-kown encoding name.

Also, the limitation "file content must not start with 0xfeff" doesn't seem
to be very serious, considering that most often CSV files start with a header
and having a value in a header starting with 0xfeff seems a recipe for disaster
don the road anyway.

closes #15